requirements.txt
flask
Flask-CORS
flask_sqlalchemy